backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ
channels==4.0.0
channels-redis==4.1.0
graphene-django==3.0.0
django-graphql-jwt==0.3.4
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
python-decouple==3.8
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
scikit-learn==1.3.2
huggingface-hub==0.19.4
tokenizers==0.15.0
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
